Primary App Server Morning Refresh - PowerShell
This script has a list of services in the order they need to be brought online to meet any dependency requirements. It offline's all ApplicationX services, and scans for process's tied to those services that may still be running after a 2 minute timer. If any process's are found to be running, it issues a kill command and rechecks for running process's. If any are found to be running still, it will wait for them to close before moving on. It then brings the specified services back online and performs an IISreset on 4 front end servers.